requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
pandas==2.2.2
scikit-learn==1.5.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-multipart==0.0.9
numpy==1.26.4
torch==2.2.0